rearranging logic
