+++ watchfiles/__init__.py
@@ -1,5 +1,6 @@
 from .filters import BaseFilter, DefaultFilter, PythonFilter
-from .main import Change, arun_process, awatch, run_process, watch
+from .main import Change, awatch, watch
+from .run import arun_process, run_process
 from .version import VERSION
 
 __version__ = VERSION
